Create README.md
The Google sheet is a list of students doing internships in companies, and includes the following columns:
A: Name of the student
B: email of the tutor in the company
C: Finish date of the internship
D: Control cell. 0 by default. Automatically equals 1 when the current date is > C. If >=1, it Increases +1 every time the Script is run, to show how many times the email has been sent to an specific company tutor.
E: Manual control cell: